# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  main.CompleteCheckoutPayload:
    properties:
      notes:
        type: string
      payment_method_id:
        type: integer
      session_id:
        type: string
      shipping_address_id:
        type: string
      shipping_method_id:
        type: integer
    required:
    - payment_method_id
    - session_id
    - shipping_address_id
    - shipping_method_id
    type: object
...
  main.StartCheckoutPayload:
    properties:
      cart_store_id:
        items:
          type: string
        type: array
    required:
    - cart_store_id
    type: object
...
  store.CartItem:
    properties:
      cart_id:
        type: integer
      cart_store_id:
        description: Nullable di database
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        allOf:
        - $ref: '#/definitions/store.Product'
        description: Relasi
      product_id:
        type: integer
      quantity:
        type: integer
      updated_at:
        type: string
    type: object
  store.CartStores:
    properties:
      cart_id:
        type: integer
      created_at:
        type: string
      id:
        type: string
      items:
        items:
          $ref: '#/definitions/store.CartItem'
        type: array
      toko:
        allOf:
        - $ref: '#/definitions/store.Toko'
        description: Relasi
      toko_id:
        type: integer
    type: object
  store.Category:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
  store.CheckoutSession:
    properties:
      cart_store:
        items:
          $ref: '#/definitions/store.CartStores'
        type: array
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: string
      notes:
        type: string
      payment_method:
        $ref: '#/definitions/store.PaymentMethod'
      session_id:
        type: string
      shipping_address:
        $ref: '#/definitions/store.ShippingAddresses'
      shipping_method:
        $ref: '#/definitions/store.ShippingMethod'
      total_price:
        type: number
      user_id:
        type: integer
    type: object
...
  store.Order:
    properties:
      created_at:
        type: string
      final_price:
        type: number
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/store.OrderItem'
        type: array
      notes:
        type: string
      order_number:
        type: string
      payment_method:
        $ref: '#/definitions/store.PaymentMethod'
      payment_method_id:
        type: integer
      shipping_addresses:
        $ref: '#/definitions/store.ShippingAddresses'
      shipping_addresses_id:
        type: string
      shipping_cost:
        type: number
      shipping_method:
        $ref: '#/definitions/store.ShippingMethod'
      shipping_method_id:
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/store.OrderStatus'
        description: Relasi
      status_id:
        type: integer
      total_price:
        type: number
      tracking:
        items:
          $ref: '#/definitions/store.OrderTracking'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  store.OrderItem:
    properties:
      created_at:
        type: string
      discount:
        type: number
      discount_price:
        type: number
      id:
        type: integer
      order_id:
        type: integer
      price:
        type: number
      product:
        allOf:
        - $ref: '#/definitions/store.Product'
        description: Relasi
      product_id:
        type: integer
      quantity:
        type: integer
      subtotal:
        type: number
      toko:
        $ref: '#/definitions/store.Toko'
      toko_id:
        type: integer
    type: object
  store.OrderStatus:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  store.OrderTracking:
    properties:
      created_at:
        type: string
      id:
        type: integer
      notes:
        type: string
      order_id:
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/store.OrderStatus'
        description: Relasi
      status_id:
        type: integer
    type: object
  store.PaymentMethod:
    properties:
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
    type: object
  store.Product:
    properties:
      category:
        $ref: '#/definitions/store.Category'
      country:
        type: string
      created_at:
        type: string
      description:
        type: string
      discount:
        type: number
      discount_price:
        type: number
      estimation:
        type: string
      id:
        type: integer
      image_urls:
        items:
          type: string
        type: array
      is_approved:
        type: boolean
      is_for_sale:
        type: boolean
      name:
        type: string
      price:
        type: number
      rating:
        type: number
      slug:
        type: string
      sold:
        type: integer
      stock:
        type: integer
      toko:
        $ref: '#/definitions/store.Toko'
      updated_at:
        type: string
      version:
        type: integer
    type: object
...
  store.ShippingAddresses:
    properties:
      address_line1:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_default:
        description: Tambahkan ini
        type: boolean
      label:
        type: string
      note_for_courier:
        type: string
      recipient_name:
        type: string
      recipient_phone:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  store.ShippingMethod:
    properties:
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
    type: object
  store.SingleUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      google_id:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      picture:
        type: string
      username:
        type: string
    type: object
  store.Toko:
    properties:
      country:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image_profile:
        type: string
      name:
        type: string
      slug:
        type: string
      user:
        $ref: '#/definitions/store.SingleUser'
      user_id:
        type: integer
    type: object
...
info:
  contact:
    email: mochammad.yogaprasetya112@gmail.com
    name: API Support
    url: http://www.myogaprasetya.my.id
  description: |-
    API ini menggunakan mekanisme rate limiter untuk membatasi jumlah permintaan yang dapat dilakukan oleh setiap klien dalam jangka waktu tertentu.
    Hal ini bertujuan untuk mencegah penyalahgunaan API, memastikan kinerja yang stabil, dan melindungi server dari potensi serangan seperti brute force atau DDoS.
    Rate limiter ini didukung oleh Redis sebagai penyimpanan sementara yang cepat dan andal untuk melacak jumlah permintaan setiap klien berdasarkan API key atau alamat IP mereka.
    Redis digunakan karena kemampuannya yang tinggi dalam menangani data secara real-time dengan latensi rendah.
    Jika batas permintaan terlampaui, klien akan menerima respons dengan kode status 429 (Too Many Requests).
  termsOfService: http://www.myogaprasetya.my.id
  title: Tokopedia API
paths:
...
  /checkout/{session_id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a checkout session
      parameters:
      - description: Checkout Session ID
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.CheckoutSession'
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get checkout session details
      tags:
      - checkout
  /checkout/complete:
    post:
      consumes:
      - application/json
      description: Complete checkout process and create order
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.CompleteCheckoutPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Order'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Complete checkout process
      tags:
      - checkout
  /checkout/start:
    post:
      consumes:
      - application/json
      description: Start checkout process and create session
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.StartCheckoutPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.CheckoutSession'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Start checkout process
      tags:
      - checkout
...
  /order:
    get:
      consumes:
      - application/json
      description: List orders with pagination and sorting
      parameters:
      - description: Number of orders to return (default 5)
        in: query
        name: limit
        type: integer
      - description: Offset for pagination (default 0)
        in: query
        name: offset
        type: integer
      - description: Sort order (asc or desc, default desc)
        in: query
        name: sort
        type: string
      - description: Search term for order details
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/store.Order'
            type: array
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: List orders for authenticated user
      tags:
      - order
  /order/{id}:
    get:
      consumes:
      - application/json
      description: Get order details by ID
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/store.Order'
        "400":
          description: Bad Request
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get order by ID
      tags:
      - order
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"